Refactor CopyButton and CopyableCell
