+++ src/frontend/src/components/barcodes/QRCode.tsx
@@ -18,11 +18,11 @@ import { useQuery } from '@tanstack/react-query';
 import QR from 'qrcode';
 import { useCallback, useEffect, useMemo, useState } from 'react';
 
+import { CopyButton } from '@lib/components/CopyButton';
 import { ApiEndpoints } from '@lib/enums/ApiEndpoints';
 import { apiUrl } from '@lib/functions/Api';
 import { api } from '../../App';
 import { useGlobalSettingsState } from '../../states/SettingsStates';
-import { CopyButton } from '../buttons/CopyButton';
 import type { QrCodeType } from '../items/ActionDropdown';
 
 import { extractErrorMessage } from '../../functions/api';
